Actually convert Pillow image colors
This dramatically improves classification performance, as you might
expect.  We still make mistakes, but can actually finish most puzzles.
We make few enough mistakes that, if we can efficiently detect when
we've gotten the puzzle into an unwinnable state, we can just start new
puzzles and expect to finish most of them.